backend/pyproject.toml
[project]
name = "movie-suggester-ai"
version = "0.1.0"
description = "AI-powered movie suggestion API"
authors = [{name = "Movie Suggester Team"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
    "exa-py>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/main.py",
    "/tests",
    "/README.md",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
start = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto" 